@@ -14,8 +14,9 @@
     </modules>
 
     <properties>
-        <maven.compiler.source>11</maven.compiler.source>
-        <maven.compiler.target>11</maven.compiler.target>
+        <java.version>13</java.version>
+        <maven.compiler.source>${java.version}</maven.compiler.source>
+        <maven.compiler.target>${java.version}</maven.compiler.target>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
         <downloadSources>true</downloadSources>
@@ -47,7 +48,19 @@
             <id>dmulloy2-repo</id>
             <url>http://repo.dmulloy2.net/nexus/repository/public/</url>
         </repository>
+
+        <repository>
+            <id>citizens-repo</id>
+            <url>https://repo.citizensnpcs.co</url>
+        </repository>
     </repositories>
+
+    <pluginRepositories>
+        <pluginRepository>
+            <id>apache-repo</id>
+            <url>https://repository.apache.org/content/repositories/snapshots</url>
+        </pluginRepository>
+    </pluginRepositories>
 
     <dependencyManagement>
         <dependencies>
@@ -128,14 +141,14 @@
             <dependency>
                 <groupId>net.citizensnpcs</groupId>
                 <artifactId>citizensapi</artifactId>
-                <version>2.0.24-SNAPSHOT</version>
+                <version>2.0.25-SNAPSHOT</version>
                 <scope>provided</scope>
             </dependency>
 
             <dependency>
                 <groupId>net.citizensnpcs</groupId>
                 <artifactId>citizens</artifactId>
-                <version>2.0.24-SNAPSHOT</version>
+                <version>2.0.25-SNAPSHOT</version>
                 <scope>provided</scope>
             </dependency>
         </dependencies>
@@ -179,7 +192,7 @@
                 <plugin>
                     <groupId>org.apache.maven.plugins</groupId>
                     <artifactId>maven-shade-plugin</artifactId>
-                    <version>3.2.1</version>
+                    <version>3.2.2-SNAPSHOT</version>
                     <executions>
                         <execution>
                             <phase>package</phase>
