@@ -1,13 +1,13 @@
 package ru.abstractcoder.murdermystery.core.game.skin.view;
 
-import ru.abstractcoder.murdermystery.core.game.player.GamePlayer;
 import ru.abstractcoder.murdermystery.core.game.role.GameRole;
+import ru.abstractcoder.murdermystery.core.game.role.RoleHolder;
 import ru.abstractcoder.murdermystery.core.game.skin.Skin;
 
 public abstract class RoleDependedSkinView implements SkinView {
 
     @Override
-    public Skin getSkinFor(GamePlayer player) {
+    public Skin getSkinFor(RoleHolder player) {
         if (player.getRole().getType() == GameRole.Type.MURDER) {
             return getSkinForMurder();
         }
