@@ -36,7 +36,9 @@ public class LobbySidebarManager {
         sidebarTemplate = lobbyEngine.settings().getSidebarTemplate()
                 .withLineUpdater(new PlaceholderTextUpdater()
                         .prepare(LobbyPlayer.class, lobbyEngine.getPlayerResolver()::resolve)
-                        .add("{desired_role}", LobbyPlayer.class, p -> p.getPreferredRole().getName())
+                        .add("{desired_role}", LobbyPlayer.class, p ->
+                                p.isRolePreferred() ? p.getPreferredRole().getName() : "???"
+                        )
                         .add("{wins}", LobbyPlayer.class, p -> p.getStatistic().getWins())
                         .add("{defeats}", LobbyPlayer.class, p -> p.getStatistic().getDefeats())
                         .add("{rating}", LobbyPlayer.class, p -> p.getStatistic().getRating())
