@@ -3,6 +3,7 @@ package ru.abstractcoder.murdermystery.core.lobby.slotbar.click;
 import com.fasterxml.jackson.annotation.JacksonInject;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import org.bukkit.event.block.Action;
+import ru.abstractcoder.benioapi.util.Lazy;
 import ru.abstractcoder.murdermystery.core.lobby.player.LobbyPlayer;
 
 import java.util.List;
@@ -10,7 +11,7 @@ import java.util.function.Consumer;
 
 public class StandartActionClickHandler extends AbstractClickHandler {
 
-    private final Consumer<LobbyPlayer> actionConsumer;
+    private final Lazy<Consumer<LobbyPlayer>> actionConsumer;
 
     @JsonCreator
     public StandartActionClickHandler(
@@ -19,12 +20,12 @@ public class StandartActionClickHandler extends AbstractClickHandler {
             @JacksonInject StandartActionResolver standartActionResolver) {
         super(allowedActions);
 
-        actionConsumer = standartActionResolver.resolve(action);
+        actionConsumer = Lazy.create(() -> standartActionResolver.resolve(action));
     }
 
     @Override
     protected void onClick(LobbyPlayer player) {
-        actionConsumer.accept(player);
+        actionConsumer.get().accept(player);
     }
 
 }
