@@ -18,7 +18,8 @@ public class StatisticRepository {
 
     public CompletableFuture<PlayerStatistic> load(String name) {
         return queryFactory.completableQuery().query(
-                "select wins, defeats, rating from statistic where username = ?",
+                "select wins, defeats, kills, deaths, golds_picked_up, rating rating from statistic " +
+                        "where username = ?",
                 rs -> {
                     if (!rs.next()) {
                         return new PlayerStatistic();
@@ -28,7 +29,7 @@ public class StatisticRepository {
                     int defeats = rs.getInt("defeats");
                     int kills = rs.getInt("kills");
                     int deaths = rs.getInt("deaths");
-                    int goldsPickedUp = rs.getInt("goldsPickedUp");
+                    int goldsPickedUp = rs.getInt("golds_picked_up");
                     int rating = rs.getInt("rating");
 
                     return new PlayerStatistic(wins, defeats, kills, deaths, goldsPickedUp, rating);
@@ -40,7 +41,10 @@ public class StatisticRepository {
     public CompletableFuture<Void> save(String name, PlayerStatistic statistic) {
         return queryFactory.completableQuery().execute(
                 //language=MySQL
-                "insert into statistic values (?, ?, ?, ?, ?, ?, ?)",
+                "insert into statistic values (?, ?, ?, ?, ?, ?, ?) on duplicate key update " +
+                        "wins = values(wins), defeats = values(defeats), kills = values(kills), " +
+                        "deaths = values(deaths), golds_picked_up = values(golds_picked_up), " +
+                        "rating = values(rating)",
                 name, statistic.getWins(), statistic.getDefeats(),
                 statistic.getKills(), statistic.getDeaths(),
                 statistic.getGoldsPickedUp(), statistic.getRating()
